etl/requirements.txt
spotipy
lyricsgenius
psycopg2-binary
python-dotenv
requests